html/browsers/windows/browsing-context-names/browsing-context-choose-existing.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: Choose browsing context - the giving name is same as a existing browsing context's name</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="help" href="http://www.w3.org/html/wg/drafts/html/CR/browsers.html#browsing-context-names">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src="existing.html" style="display:none"></iframe>
<iframe src="existing.html" name="existWin" style="display:none"></iframe>
<script>

var t = async_test("The browsing context must be chosen if the giving name is same as its name");

window.addEventListener("message", function (e) {
  t.step(function() {
    assert_equals(e.data.name, "existWin", "The window context name should be 'topWin'.");
  });
  t.done();
}, false);

</script>
